Force synchronization before imminent history overflow
Added a synchronization mechanism that collects all events back to the queue, to properly recompute GVT. It allows to reduce the history sizes and settles the proper ordering of events. It's a means to an end and will only be called when the history buffer is close to being overflowed for any LP.
Also fixed the LFSR initialization seeds in the cores.